pyproject.toml
[tool.poetry]
name = "acrotrainer"
version = "0.1.0"
description = ""
authors = ["seth mcfeeters <seth.mcfeeters@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
django-debug-toolbar = "^3.2"
django-environ = "^0.4.5"
pytest = "^6.2.2"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
six = "^1.15.0"
whitenoise = "^5.2.0"
pytest-django = "^4.1.0"
Pillow = "^8.1.0"
django-crispy-forms = "^1.11.1"
django-allauth = "^0.44.0"
coverage = "^5.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
